print indices as ints
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@10426 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/obstacle/MeshFace.cxx
@@ -107,7 +107,7 @@ void MeshFace::finalize()
     if ((debugLevel >= 3) && (mesh != NULL)) {
       printf(":");
       for (i = 0; i < vertexCount; i++) {
-        printf(" %f", (double)((fvec3*)vertices[i] - mesh->getVertices()));
+        printf(" %i", (int)((fvec3*)vertices[i] - mesh->getVertices()));
       }
       print(std::cerr, "");
     }
@@ -139,7 +139,7 @@ void MeshFace::finalize()
       if ((debugLevel >= 3) && (mesh != NULL)) {
         printf(":");
         for (i = 0; i < vertexCount; i++) {
-          printf(" %f", (double)((fvec3*)vertices[i] - mesh->getVertices()));
+          printf(" %i", (int)((fvec3*)vertices[i] - mesh->getVertices()));
         }
 	print(std::cerr, "");
       }
@@ -158,7 +158,7 @@ void MeshFace::finalize()
       if ((debugLevel >= 3) && (mesh != NULL)) {
         printf(":");
         for (i = 0; i < vertexCount; i++) {
-          printf(" %f", (double)((fvec3*)vertices[i] - mesh->getVertices()));
+          printf(" %i", (int)((fvec3*)vertices[i] - mesh->getVertices()));
         }
 	print(std::cerr, "");
       }
